Refactor CollectionCard and CollectionSection styles; add responsive design and hover animations; update WhyUsSection and content structure
