+++ src/components/organism/CollectionSection.module.css
@@ -1,0 +1,65 @@
+.collection {
+    display: grid;
+    grid-template-columns: 1fr;
+    grid-template-rows: 100px 1fr; /* Changed to allow .collection_list to expand */
+    height: 100%; /* Ensure .collection takes up available height */
+  }
+  
+  .collection_header {
+      display: grid;
+      place-self: center;
+
+      font-size: 8rem;
+    text-align: center;
+    position: relative;
+  }
+  
+  .collection_header::after {
+      content: "";
+      position: absolute;
+      top: 0px;
+      bottom: 0;
+      right: 0px;
+      left: 0px;
+      background-image: linear-gradient(rgba(254, 252, 247, 0) 0%, rgb(254, 252, 247) 80%);
+  }
+  
+  .collection_list {
+      position: relative;
+      z-index: 99;
+     transform: translateY(-30px);
+    display: grid;
+    place-items: center;
+    grid-template-columns: repeat(auto-fit, minmax(290px, 1fr));
+    grid-auto-rows: minmax(250px, 1fr); /* Allow rows to adjust height */
+    height: 100%; /* Ensure .collection_list takes up available height */
+  }
+  
+  @media (max-width: 1200px) {
+      .collection_header {
+          font-size: 5rem;
+          grid-row: 1/2;
+      }
+    .collection_list {
+      grid-template-columns: repeat(2, 1fr);
+      grid-auto-rows: minmax(250px, 1fr); /* Allow rows to adjust height */
+      gap: 1rem;
+    }
+    .collection {
+      grid-template-rows: 50px 1fr; /* Allow .collection_list to expand */
+    }
+  }
+  
+  @media (max-width: 700px) {
+      .collection {
+      grid-template-rows: 50px 1fr; /* Allow .collection_list to expand */
+        }
+      .collection_header {
+          font-size: 2.4rem;
+      }
+    .collection_list {
+      grid-template-columns: 1fr;
+      grid-auto-rows: minmax(250px, 1fr); /* Allow rows to adjust height */
+    }
+  }
+
