+++ src/components/organism/CollectionSection.tsx
@@ -1,12 +1,12 @@
 import CollectionCard from '../atoms/CollectionCard';
 import { collection } from '../../ultils/content';
-
+import style from './CollectionSection.module.css';
 const Collection: React.FC = () => {
 
   return (
-    <section>
-      <h2>Our Collection</h2>
-      <ul>
+    <section className={style.collection}>
+      <h2 className={style.collection_header}>Our Collection</h2>
+      <ul className={style.collection_list}>
         {collection.map((item) => (
           <CollectionCard
             key={item.title}
